requirements.txt
Django
djangorestframework
djangorestframework-jwt
django-cors-headers
django-rest-auth
Pillow
requests
transitions
ipython
